Add very basic tests for model validation
